+++ doc/changelog.rst
@@ -22,6 +22,7 @@ Next Release
 - Bug Fixes:
 
   - :issue_gh:`20` program icon in task bar missing since version 0.13.x
+  - :issue_ghe:`721` Correct execution history logging
   - :issue_ghe:`726` State with self-transition cannot be substituted
   - fixes bug of "locked" global variable during multithreading access
   - use a safe loader for GUI config file